separate out instructions into different component
