+++ static/js/src/advantage/attach/components/AttachInstructions/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./AttachInstructions";
